ml-service/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
requests
pillow
pydantic
pymongo
redis